Update smi2srt.sh
변환할 파일이 없다면 skip

+++ smi2srt.sh
@@ -3,11 +3,9 @@
 # convert smi to ko.srt
 # if <body> include : smi2srt by "George Shuklin"
 # if <body> doesn`t include : smi2srt by axfree npm
-# sudo npm install smi2srt -g
-
 targetdir='/volume1/video/해외영화'
 targetdir2='/volume1/video/4K UHD (2160p)'
-smidir='/volume1/homes/user/script/smi2srt'
+smidir='/volume1/homes/yklovett/script/smi2srt'
 
 find "$targetdir" "$targetdir2" -name "*.smi" | grep -v "@eaDir" | grep -v "#recycle" | while read filename
     do
@@ -24,35 +22,46 @@ find "$targetdir" "$targetdir2" -name "*.smi" | grep -v "@eaDir" | grep -v "#rec
 	done
 
 		# convert include <body> in subtitles
+if [ -f "$smidir/grep_list" ]
+then
 	while read oldfile
-		do
-			newfile=${oldfile%smi}ko.srt
-			    if [ -f "$newfile" ]
-			        then
-				        echo "$oldfile" "Skip!!"
-				    else
-				        $smidir/smi2srt "$oldfile" "$newfile"
-				        echo "$oldfile => $newfile" converted
-						# remove origin file
-				        rm -f "$oldfile"
-			    fi
-		done < $smidir/grep_list
-
-       # convert doesn`t include <body> in subtitles
-	while read oldfile2
-		do
-			newfile2=${oldfile2%smi}ko.srt
-				if [ -f "$newfile2" ]
-					then
-						echo "$oldfile2" "Skip!!"
-					else
-						/usr/local/bin/smi2srt -n "$oldfile2"
-						echo "$oldfile2 => $newfile2" converted
-						# remove origin file
-						rm -f "$oldfile2"
-				fi
-		done < $smidir/grep_notlist
-
+	do
+		newfile=${oldfile%smi}ko.srt
+		    if [ -f "$newfile" ]
+	        then
+		        echo "$oldfile" "Skip!!"
+		    else
+		        $smidir/smi2srt "$oldfile" "$newfile"
+		        echo "$oldfile => $newfile" converted
+				# remove origin file
+		        rm -f "$oldfile"
+		    fi
+	done < $smidir/grep_list
 sleep 1
-# remove smi file list
-rm -f $smidir/grep_*list
+rm -f $smidir/grep_list
+
+else
+	echo "### no subtitles to convert (include body)"
+fi
+       # convert doesn`t include <body> in subtitles
+if [ -f "$smidir/grep_notlist" ]
+then
+	while read oldfile2
+	do
+		newfile2=${oldfile2%smi}ko.srt
+				if [ -f "$newfile2" ]
+				then
+					echo "$oldfile2" "Skip!!"
+				else
+					/usr/local/bin/smi2srt -n "$oldfile2"
+					echo "$oldfile2 => $newfile2" converted
+					# remove origin file
+					rm -f "$oldfile2"
+				fi
+	done < $smidir/grep_notlist
+sleep 1
+rm -f $smidir/grep_notlist
+
+else
+	echo "### no subtitles to convert (not include body)"
+fi
